+++ declarations.d.ts
@@ -1,2 +1,3 @@
-declare module '@carbon/icons-react';
-declare module 'wicg-inert';
+declare module "@carbon/icons-react";
+declare module "carbon-components-react/es/index";
+declare module "wicg-inert";