+++ src/components/SimpleTable/index.tsx
@@ -1,4 +1,0 @@
-import{ Table, TableHead, TableRow, TableBody, TableCell, TableHeader } from "carbon-components-react";
-import TableContainer from "./TableContainer";
-
-export { TableContainer, Table, TableHead, TableRow, TableBody, TableCell, TableHeader };