+++ src/components/Table/index.tsx
@@ -1,0 +1,14 @@
+export * from "./TableContainer";
+export * from "./Table";
+export * from "./TableRow";
+export * from "./TableExpandRow";
+export * from "./TableExpandHeader";
+export * from "./TableCell";
+
+// TODO, we may wrap to add own classes (currently not necessary)
+export {
+    TableHead,
+    TableBody,
+    TableExpandedRow,
+    TableHeader,
+} from "carbon-components-react";