+++ src/components/index.ts
@@ -1,0 +1,10 @@
+export * from "./Badge/Badge";
+export * from "./Depiction/Depiction";
+export * from "./Icon";
+export * from "./ProgressBar";
+export * from "./Table";
+export * from "./Tag";
+export * from "./Tree/Tree";
+export * from "./Typography";
+
+export * from "./interfaces";